Add CPU suggestions and report table utilities for resource dashboard
- Introduced `dashboard_cpu_suggestions.py` to provide CPU recommendations based on test runs, including logic for handling timeouts and suite size policies.
- Added `dashboard_report_table.py` for generating an HTML report table from test results, allowing for filtering and detailed metrics display.
- Updated `tests_resource_dashboard.py` to integrate the new CPU suggestion and report table functionalities, enhancing the overall resource dashboard capabilities.
