+++ .github/scripts/tests/utils/tests_resource_dashboard.py
@@ -23,9 +23,13 @@ from typing import Any, Optional
 # Same-dir library for ya.make REQUIREMENTS (reserved cpu/ram)
 if __name__ != "__main__":
     from . import ya_make_requirements
+    from .dashboard_cpu_suggestions import build_cpu_suggestions
+    from .dashboard_report_table import build_report_table_html
 else:
     sys.path.insert(0, str(Path(__file__).resolve().parent))
     import ya_make_requirements
+    from dashboard_cpu_suggestions import build_cpu_suggestions
+    from dashboard_report_table import build_report_table_html
 
 # Supports both:
 #   [3/10] chunk
@@ -623,202 +627,6 @@ def _downsample_active_series(
         ys[bisect.bisect_right(xs, t) - 1] if xs[0] <= t else 0.0 for t in xs_sampled
     ]
     return xs_sampled, ys_sampled
-
-
-def _round_cpu_tier(cores: float) -> int:
-    """Round up to runner-friendly tier: 1, 2, 4, 8, 16."""
-    if cores <= 1:
-        return 1
-    if cores <= 2:
-        return 2
-    if cores <= 4:
-        return 4
-    if cores <= 8:
-        return 8
-    return 16
-
-
-def _next_cpu_tier(cpu: int) -> int:
-    """Move to the next runner tier: 1->2->4->8->16."""
-    if cpu <= 1:
-        return 2
-    if cpu <= 2:
-        return 4
-    if cpu <= 4:
-        return 8
-    if cpu <= 8:
-        return 16
-    return 16
-
-
-def build_cpu_suggestions(
-    runs: list[dict[str, Any]],
-    requirements_cache: Optional[dict[str, dict[str, Any]]] = None,
-    report_status_by_suite: Optional[dict[str, dict[str, dict[str, int]]]] = None,
-    maximize_reqs_for_timeout_tests: bool = False,
-) -> list[dict[str, Any]]:
-    """
-    Per-suite stats from runs (cpu_sec_report, dur_us) to suggest cpuN for the runner.
-    Returns list of dicts: suite_path, chunks_count, median_cores, p95_cores,
-    recommended_cpu (tier 1/2/4/8/16 or "all"), total_cpu_sec, total_ram_gb, total_dur_sec, has_synthetic.
-    """
-    by_suite: dict[str, list[float]] = defaultdict(list)
-    by_suite_cpu: dict[str, float] = defaultdict(float)
-    by_suite_ram_kb: dict[str, float] = defaultdict(float)
-    by_suite_dur: dict[str, float] = defaultdict(float)
-    by_suite_synthetic: dict[str, bool] = defaultdict(bool)
-    by_suite_errors: dict[str, int] = defaultdict(int)
-    by_suite_timeouts: dict[str, int] = defaultdict(int)
-    by_suite_muted: dict[str, int] = defaultdict(int)
-    by_suite_muted_timeouts: dict[str, int] = defaultdict(int)
-    for r in runs:
-        suite = str(r.get("suite_path", ""))
-        if not suite:
-            continue
-        cpu = float(r.get("cpu_sec_report", 0.0) or 0.0)
-        ram_kb = float(r.get("ram_kb_report", 0.0) or 0.0)
-        dur_s = float(r.get("dur_us", 0) or 0) / 1_000_000.0
-        by_suite_cpu[suite] += cpu
-        by_suite_ram_kb[suite] += ram_kb
-        by_suite_dur[suite] += dur_s
-        if r.get("synthetic_metrics"):
-            by_suite_synthetic[suite] = True
-        status = str(r.get("status", "") or "").upper()
-        error_type = str(r.get("error_type", "") or "").upper()
-        is_timeout = error_type == "TIMEOUT" or ("TIMEOUT" in status)
-        is_muted = bool(r.get("is_muted")) or status == "MUTE"
-        is_failedish = status in {"FAILED", "ERROR", "INTERNAL"}
-        if is_timeout:
-            by_suite_timeouts[suite] += 1
-        if is_muted:
-            by_suite_muted[suite] += 1
-        if is_timeout and is_muted:
-            by_suite_muted_timeouts[suite] += 1
-        if is_failedish and not is_timeout and not is_muted:
-            by_suite_errors[suite] += 1
-        if dur_s > 0:
-            by_suite[suite].append(cpu / dur_s)
-    all_suites = sorted(set(by_suite_cpu.keys()) | set((report_status_by_suite or {}).keys()) | set((requirements_cache or {}).keys()))
-    out: list[dict[str, Any]] = []
-    for suite in all_suites:
-        cores_list = by_suite[suite]
-        if not cores_list:
-            median_c = 0.0
-            p95_c = 0.0
-        else:
-            sorted_cores = sorted(cores_list)
-            n = len(sorted_cores)
-            median_c = sorted_cores[(n - 1) // 2] if n else 0.0
-            idx95 = min(int(0.95 * n + 0.5), n - 1) if n else 0
-            p95_c = sorted_cores[idx95] if n else 0.0
-        if report_status_by_suite and suite in report_status_by_suite:
-            chunk_status = report_status_by_suite[suite].get("chunks", _status_bucket())
-            test_status = report_status_by_suite[suite].get("tests", _status_bucket())
-        else:
-            chunk_status = {
-                "errors": by_suite_errors[suite],
-                "timeouts": by_suite_timeouts[suite],
-                "muted": by_suite_muted[suite],
-                "muted_timeouts": by_suite_muted_timeouts[suite],
-                "fails_total": by_suite_errors[suite] + by_suite_timeouts[suite],
-            }
-            test_status = _status_bucket()
-
-        timeout_tests_count = int(test_status.get("timeouts", 0) or 0)
-        base_recommended = _round_cpu_tier(p95_c)
-        timeout_boost_applied = timeout_tests_count > 0
-        recommended = _next_cpu_tier(base_recommended) if timeout_boost_applied else base_recommended
-        req = (requirements_cache or {}).get(suite, {})
-        ya_cpu = req.get("cpu_cores")
-        ya_ram = req.get("ram_gb")
-        ya_size = req.get("size")
-        if not ya_size:
-            ya_size = "SMALL"
-        # Policy: MEDIUM suites must not request more than cpu:4.
-        medium_cap_applied = str(ya_size or "").upper() == "MEDIUM" and recommended > 4
-        if medium_cap_applied:
-            recommended = 4
-        timeout_max_policy_applied = bool(maximize_reqs_for_timeout_tests and timeout_tests_count > 0)
-        size_u = str(ya_size or "").upper()
-        timeout_max_value: Any = None
-        if timeout_max_policy_applied:
-            if size_u == "SMALL":
-                timeout_max_value = 1
-            elif size_u == "MEDIUM":
-                timeout_max_value = 4
-            elif size_u == "LARGE":
-                timeout_max_value = "all"
-            else:
-                timeout_max_value = "all"
-        recommended_req: Any = timeout_max_value if timeout_max_policy_applied else recommended
-        explain_parts = [
-            f"p95_cores={p95_c:.3f}",
-            f"base_tier={base_recommended}",
-        ]
-        if timeout_boost_applied:
-            explain_parts.append(f"test_timeouts={timeout_tests_count} -> next_tier={_next_cpu_tier(base_recommended)}")
-        if timeout_max_policy_applied:
-            explain_parts.append(f"maximize_reqs_for_timeout_tests(size={size_u or 'UNKNOWN'}) -> {timeout_max_value}")
-        if medium_cap_applied:
-            explain_parts.append("MEDIUM cap -> 4")
-        explain_parts.append(f"final={recommended_req}")
-        recommended_cpu_explain = "; ".join(explain_parts)
-        if recommended_req == "all":
-            if ya_cpu is None:
-                cpu_action = "set"
-            else:
-                cpu_action = "raise"
-        else:
-            recommended_num = int(recommended_req)
-            if ya_cpu is None:
-                # In ya.make implicit default is cpu:1, so do not suggest explicit set(cpu:1).
-                cpu_action = "ok" if recommended_num <= 1 else "set"
-            else:
-                ya_cpu_i = int(ya_cpu)
-                if recommended_num > ya_cpu_i:
-                    cpu_action = "raise"
-                elif recommended_num < ya_cpu_i:
-                    cpu_action = "lower"
-                else:
-                    cpu_action = "ok"
-        out.append({
-            "suite_path": suite,
-            "chunks_count": len(cores_list),
-            "median_cores": round(median_c, 3),
-            "p95_cores": round(p95_c, 3),
-            "recommended_cpu": recommended_req,
-            "recommended_cpu_base": base_recommended,
-            "recommended_cpu_timeout_boost": timeout_boost_applied,
-            "recommended_cpu_timeout_max_policy_applied": timeout_max_policy_applied,
-            "recommended_cpu_timeout_max_policy_value": timeout_max_value,
-            "recommended_cpu_medium_cap_applied": medium_cap_applied,
-            "recommended_cpu_explain": recommended_cpu_explain,
-            "total_cpu_sec": round(by_suite_cpu[suite], 2),
-            "total_ram_gb": round(by_suite_ram_kb[suite] / (1024.0 * 1024.0), 3),
-            "total_dur_sec": round(by_suite_dur[suite], 2),
-            "has_synthetic": by_suite_synthetic.get(suite, False),
-            "ya_cpu_cores": ya_cpu,
-            "ya_ram_gb": ya_ram,
-            "ya_size": ya_size,
-            "chunk_errors": chunk_status["errors"],
-            "chunk_timeouts": chunk_status["timeouts"],
-            "chunk_muted": chunk_status["muted"],
-            "chunk_muted_timeouts": chunk_status["muted_timeouts"],
-            "chunk_fails_total": chunk_status["fails_total"],
-            "test_errors": test_status["errors"],
-            "test_timeouts": test_status["timeouts"],
-            "test_muted": test_status["muted"],
-            "test_muted_timeouts": test_status["muted_timeouts"],
-            "test_fails_total": test_status["fails_total"],
-            # Legacy aliases kept in output schema for stability.
-            "errors": chunk_status["errors"],
-            "timeouts": chunk_status["timeouts"],
-            "muted": chunk_status["muted"],
-            "muted_timeouts": chunk_status["muted_timeouts"],
-            "fails_total": chunk_status["fails_total"],
-            "cpu_action": cpu_action,
-        })
-    return out
 
 
 def build_html_dashboard(
@@ -2030,293 +1838,6 @@ def build_html_dashboard(
       suiteSearchEl.addEventListener('input', applySuiteSearch);
     }}
     applySuiteSearch();
-  </script>
-</body>
-</html>
-"""
-    out_html.write_text(html, encoding="utf-8")
-
-
-def build_report_table_html(report_path: Path, out_html: Path, suite_filter: Optional[str]) -> None:
-    report = json.loads(report_path.read_text(encoding="utf-8", errors="replace"))
-    results = report.get("results", []) if isinstance(report, dict) else []
-    rows: list[dict[str, Any]] = []
-    for item in results:
-        if not isinstance(item, dict):
-            continue
-        if item.get("type") != "test":
-            continue
-        suite_raw = str(item.get("path", "") or "")
-        if not suite_raw:
-            continue
-        suite = normalize_suite_path(suite_raw)
-        if suite_filter and suite != suite_filter:
-            continue
-
-        metrics = item.get("metrics") if isinstance(item.get("metrics"), dict) else {}
-        sub = str(item.get("subtest_name", "") or "")
-        chunk_idx: Optional[int] = None
-        m_idx = CHUNK_FROM_SUBTEST_RE.search(sub)
-        if m_idx:
-            chunk_idx = int(m_idx.group(1))
-        elif CHUNK_SOLE_RE.search(sub) or CHUNK_BRACKET_ONLY_RE.search(sub):
-            chunk_idx = 0
-        chunk_group = chunk_group_from_subtest(sub)
-
-        rows.append(
-            {
-                "suite_path": suite,
-                "suite_path_raw": suite_raw,
-                "test_name": str(item.get("name", "") or ""),
-                "subtest_name": sub,
-                "status": str(item.get("status", "") or ""),
-                "error_type": str(item.get("error_type", "") or ""),
-                "is_muted": bool(item.get("is_muted") or item.get("muted")),
-                "chunk": bool(item.get("chunk")),
-                "chunk_idx": chunk_idx,
-                "chunk_group": chunk_group,
-                "duration_sec": float(item.get("duration") or 0.0),
-                "cpu_sec": cpu_seconds(metrics),
-                "ram_kb": ram_kb(metrics),
-                "ru_utime": metrics.get("ru_utime"),
-                "ru_stime": metrics.get("ru_stime"),
-                "ru_maxrss": metrics.get("ru_maxrss"),
-                "ru_rss": metrics.get("ru_rss"),
-                "wall_time": metrics.get("wall_time"),
-                "id": item.get("id"),
-                "hid": item.get("hid"),
-                "size": item.get("size"),
-                "tags": ", ".join(item.get("tags", [])) if isinstance(item.get("tags"), list) else "",
-                "metrics_json": json.dumps(metrics, ensure_ascii=False, sort_keys=True),
-            }
-        )
-
-    payload = {
-        "suite_filter": suite_filter,
-        "report_path": str(report_path),
-        "rows_count": len(rows),
-        "rows": rows,
-    }
-    html = f"""<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <title>Report table: suite/chunk/test</title>
-  <style>
-    body {{ font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Arial, sans-serif; margin: 12px; }}
-    .toolbar {{ position: sticky; top: 0; background: #fff; z-index: 3; border-bottom: 1px solid #eee; padding: 8px 0; display: flex; gap: 8px; align-items: center; flex-wrap: wrap; }}
-    .toolbar input, .toolbar select {{ padding: 4px 8px; }}
-    .table-wrap {{ overflow-x: auto; border: 1px solid #eee; border-radius: 6px; }}
-    table {{ width: 100%; min-width: 2650px; border-collapse: collapse; font-size: 12px; table-layout: fixed; }}
-    th, td {{ border: 1px solid #eee; padding: 4px 6px; vertical-align: top; }}
-    th {{ position: sticky; top: 48px; background: #fafafa; z-index: 2; cursor: pointer; user-select: none; }}
-    td pre {{ margin: 0; white-space: pre-wrap; word-break: break-word; max-width: 620px; }}
-    #tbody tr {{ content-visibility: auto; contain-intrinsic-size: 28px; }}
-    th:nth-child(1), td:nth-child(1) {{ width: 320px; }}
-    th:nth-child(2), td:nth-child(2) {{ width: 70px; }}
-    th:nth-child(3), td:nth-child(3) {{ width: 130px; }}
-    th:nth-child(4), td:nth-child(4) {{ width: 70px; }}
-    th:nth-child(5), td:nth-child(5) {{ width: 220px; }}
-    th:nth-child(6), td:nth-child(6) {{ width: 260px; }}
-    th:nth-child(22), td:nth-child(22) {{ width: 130px; }}
-    .mono {{ font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace; }}
-    .muted {{ color: #6a737d; }}
-    .ok {{ color: #155724; }}
-    .fail {{ color: #721c24; }}
-    .sidepanel {{ position: fixed; top: 0; right: 0; width: min(52vw, 820px); height: 100vh; background: #fff; border-left: 1px solid #ddd; box-shadow: -8px 0 24px rgba(0,0,0,0.08); z-index: 20; display: none; }}
-    .sidepanel .head {{ display: flex; justify-content: space-between; align-items: center; padding: 10px 12px; border-bottom: 1px solid #eee; }}
-    .sidepanel .body {{ padding: 10px 12px; height: calc(100vh - 56px); overflow: auto; }}
-  </style>
-</head>
-<body>
-  <h2>Report table: suite/chunk/test</h2>
-  <div class="muted">suite filter: {suite_filter or 'ALL SUITES'} | rows: <span id="rowsCount"></span></div>
-  <div class="toolbar">
-    <label>Search:</label>
-    <input id="q" type="text" placeholder="suite/test/subtest/status/tags" style="min-width: 360px;" />
-    <label>Status:</label>
-    <select id="statusSel">
-      <option value="">all</option>
-    </select>
-    <label>Mode:</label>
-    <select id="chunkSel">
-      <option value="">all</option>
-      <option value="chunk">chunk=true</option>
-      <option value="test">chunk=false</option>
-    </select>
-    <button type="button" onclick="clearFilters()">Clear</button>
-  </div>
-  <div class="table-wrap">
-    <table id="tbl">
-      <thead id="thead"></thead>
-      <tbody id="tbody"></tbody>
-    </table>
-  </div>
-  <div id="metricsPanel" class="sidepanel" role="dialog" aria-modal="true">
-    <div class="head">
-      <b id="metricsPanelTitle">metrics</b>
-      <button type="button" onclick="closeMetricsPanel()">Close</button>
-    </div>
-    <div class="body">
-      <pre id="metricsPanelBody" class="mono"></pre>
-    </div>
-  </div>
-  <script>
-    const data = {json.dumps(payload, ensure_ascii=False)};
-    const cols = [
-      ['suite_path', 'suite_path'],
-      ['chunk', 'chunk'],
-      ['chunk_group', 'chunk_group'],
-      ['chunk_idx', 'chunk_idx'],
-      ['test_name', 'test_name'],
-      ['subtest_name', 'subtest_name'],
-      ['status', 'status'],
-      ['error_type', 'error_type'],
-      ['is_muted', 'is_muted'],
-      ['duration_sec', 'duration_sec'],
-      ['cpu_sec', 'cpu_sec'],
-      ['ram_kb', 'ram_kb'],
-      ['ru_utime', 'ru_utime'],
-      ['ru_stime', 'ru_stime'],
-      ['ru_maxrss', 'ru_maxrss'],
-      ['ru_rss', 'ru_rss'],
-      ['wall_time', 'wall_time'],
-      ['size', 'size'],
-      ['id', 'id'],
-      ['hid', 'hid'],
-      ['tags', 'tags'],
-      ['metrics_json', 'metrics'],
-    ];
-
-    let sortCol = 'suite_path';
-    let sortAsc = true;
-
-    function valueForSort(v) {{
-      if (v === null || v === undefined) return '';
-      if (typeof v === 'boolean') return v ? 1 : 0;
-      const n = Number(v);
-      if (!Number.isNaN(n) && String(v).trim() !== '') return n;
-      return String(v).toLowerCase();
-    }}
-
-    function esc(s) {{
-      return String(s ?? '')
-        .replaceAll('&', '&amp;')
-        .replaceAll('<', '&lt;')
-        .replaceAll('>', '&gt;')
-        .replaceAll('"', '&quot;');
-    }}
-
-    function render() {{
-      const q = (document.getElementById('q').value || '').toLowerCase().trim();
-      const st = document.getElementById('statusSel').value;
-      const mode = document.getElementById('chunkSel').value;
-
-      const filtered = data.rows.map((r, idx) => ({{r, idx}})).filter(({{r}}) => {{
-        if (st && String(r.status || '') !== st) return false;
-        if (mode === 'chunk' && !r.chunk) return false;
-        if (mode === 'test' && r.chunk) return false;
-        if (!q) return true;
-        const hay = [
-          r.suite_path, r.test_name, r.subtest_name, r.status, r.error_type, r.tags, r.chunk_group, r.id, r.hid
-        ].map(x => String(x || '').toLowerCase()).join(' ');
-        return hay.includes(q);
-      }});
-
-      filtered.sort((a, b) => {{
-        const av = valueForSort(a.r[sortCol]);
-        const bv = valueForSort(b.r[sortCol]);
-        if (av < bv) return sortAsc ? -1 : 1;
-        if (av > bv) return sortAsc ? 1 : -1;
-        return 0;
-      }});
-
-      const head = '<tr>' + cols.map(([k, title]) => {{
-        const marker = sortCol === k ? (sortAsc ? ' ▲' : ' ▼') : '';
-        return '<th data-col=\"' + k + '\">' + esc(title) + marker + '</th>';
-      }}).join('') + '</tr>';
-      document.getElementById('thead').innerHTML = head;
-      document.querySelectorAll('#thead th').forEach(th => th.addEventListener('click', () => {{
-        const c = th.getAttribute('data-col');
-        if (sortCol === c) sortAsc = !sortAsc;
-        else {{ sortCol = c; sortAsc = true; }}
-        render();
-      }}));
-
-      const body = filtered.map(({{r, idx}}) => {{
-        const statusCls = /^(OK|PASS)$/i.test(String(r.status || '')) ? 'ok' : (/^(FAILED|ERROR|TIMEOUT|INTERNAL|MUTE)$/i.test(String(r.status || '')) ? 'fail' : '');
-        const metricsCell = '<button type=\"button\" class=\"metrics-btn\" data-idx=\"' + String(idx) + '\">open metrics</button>';
-        const vals = {{
-          suite_path: esc(r.suite_path),
-          chunk: r.chunk ? 'true' : 'false',
-          chunk_group: esc(r.chunk_group || ''),
-          chunk_idx: r.chunk_idx == null ? '' : String(r.chunk_idx),
-          test_name: esc(r.test_name || ''),
-          subtest_name: esc(r.subtest_name || ''),
-          status: '<span class=\"' + statusCls + '\">' + esc(r.status || '') + '</span>',
-          error_type: esc(r.error_type || ''),
-          is_muted: r.is_muted ? 'true' : 'false',
-          duration_sec: Number(r.duration_sec || 0).toFixed(3),
-          cpu_sec: Number(r.cpu_sec || 0).toFixed(6),
-          ram_kb: Number(r.ram_kb || 0).toFixed(0),
-          ru_utime: esc(r.ru_utime ?? ''),
-          ru_stime: esc(r.ru_stime ?? ''),
-          ru_maxrss: esc(r.ru_maxrss ?? ''),
-          ru_rss: esc(r.ru_rss ?? ''),
-          wall_time: esc(r.wall_time ?? ''),
-          size: esc(r.size ?? ''),
-          id: '<span class=\"mono\">' + esc(r.id ?? '') + '</span>',
-          hid: '<span class=\"mono\">' + esc(r.hid ?? '') + '</span>',
-          tags: esc(r.tags || ''),
-          metrics_json: metricsCell,
-        }};
-        return '<tr>' + cols.map(([k]) => '<td>' + (vals[k] ?? '') + '</td>').join('') + '</tr>';
-      }}).join('');
-
-      document.getElementById('tbody').innerHTML = body;
-      document.querySelectorAll('.metrics-btn').forEach(btn => btn.addEventListener('click', () => {{
-        const idx = Number(btn.getAttribute('data-idx'));
-        openMetricsPanel(idx);
-      }}));
-      document.getElementById('rowsCount').textContent = String(filtered.length);
-    }}
-
-    function openMetricsPanel(idx) {{
-      const row = data.rows[idx];
-      if (!row) return;
-      const title = (row.suite_path || '') + ' | ' + (row.test_name || '') + (row.subtest_name ? (' | ' + row.subtest_name) : '');
-      const body = row.metrics_json || '{{}}';
-      const panel = document.getElementById('metricsPanel');
-      document.getElementById('metricsPanelTitle').textContent = title;
-      document.getElementById('metricsPanelBody').textContent = body;
-      panel.style.display = 'block';
-    }}
-
-    function closeMetricsPanel() {{
-      const panel = document.getElementById('metricsPanel');
-      panel.style.display = 'none';
-    }}
-
-    function clearFilters() {{
-      document.getElementById('q').value = '';
-      document.getElementById('statusSel').value = '';
-      document.getElementById('chunkSel').value = '';
-      render();
-    }}
-
-    const statuses = Array.from(new Set(data.rows.map(r => String(r.status || '')).filter(Boolean))).sort();
-    const stSel = document.getElementById('statusSel');
-    statuses.forEach(s => {{
-      const opt = document.createElement('option');
-      opt.value = s;
-      opt.textContent = s;
-      stSel.appendChild(opt);
-    }});
-    document.getElementById('q').addEventListener('input', render);
-    document.getElementById('statusSel').addEventListener('change', render);
-    document.getElementById('chunkSel').addEventListener('change', render);
-    render();
   </script>
 </body>
 </html>
